+++ src/lib/ProjectM36/Base.hs
@@ -24,6 +24,7 @@ import Data.Time.Calendar (Day)
 import Data.Typeable
 import Data.ByteString (ByteString)
 import qualified Data.List.NonEmpty as NE
+import Data.Vector.Instances ()
 
 type StringType = Text
 
@@ -39,8 +40,9 @@ data Atom = IntegerAtom Integer |
             ByteStringAtom ByteString |
             BoolAtom Bool |
             RelationAtom Relation |
+            RelationalExprAtom RelationalExpr | --used for returning inc deps
             ConstructedAtom DataConstructorName AtomType [Atom]
-            deriving (Eq, Show, Typeable, NFData, Generic)
+            deriving (Eq, Show, Typeable, NFData, Generic, Read)
                      
 instance Hashable Atom where                     
   hashWithSalt salt (ConstructedAtom dConsName _ atoms) = salt `hashWithSalt` atoms
@@ -54,6 +56,7 @@ instance Hashable Atom where
   hashWithSalt salt (ByteStringAtom bs) = salt `hashWithSalt` bs
   hashWithSalt salt (BoolAtom b) = salt `hashWithSalt` b
   hashWithSalt salt (RelationAtom r) = salt `hashWithSalt` r
+  hashWithSalt salt (RelationalExprAtom re) = salt `hashWithSalt` re
 
 -- I suspect the definition of ConstructedAtomType with its name alone is insufficient to disambiguate the cases; for example, one could create a type named X, remove a type named X, and re-add it using different constructors. However, as long as requests are served from only one DatabaseContext at-a-time, the type name is unambiguous. This will become a problem for time-travel, however.
 -- | The AtomType uniquely identifies the type of a atom.
@@ -67,13 +70,15 @@ data AtomType = IntAtomType |
                 BoolAtomType |
                 RelationAtomType Attributes |
                 ConstructedAtomType TypeConstructorName TypeVarMap |
+                RelationalExprAtomType |
                 TypeVariableType TypeVarName
                 --wildcard used in Atom Functions and tuples for data constructors which don't provide all arguments to the type constructor
-              deriving (Eq, NFData, Generic, Show)
+              deriving (Eq, NFData, Generic, Show, Read, Hashable)
 
 instance Ord AtomType where
   compare = undefined
-  
+
+-- this should probably be an ordered dictionary in order to be able to round-trip these arguments  
 type TypeVarMap = M.Map TypeVarName AtomType
 
 instance Hashable TypeVarMap where 
@@ -88,7 +93,7 @@ isRelationAtomType _ = False
 type AttributeName = StringType
 
 -- | A relation's type is composed of attribute names and types.
-data Attribute = Attribute AttributeName AtomType deriving (Eq, Show, Generic, NFData)
+data Attribute = Attribute AttributeName AtomType deriving (Eq, Show, Read, Generic, NFData)
 
 instance Hashable Attribute where
   hashWithSalt salt (Attribute attrName _) = hashWithSalt salt attrName
@@ -106,7 +111,10 @@ sortedAttributesIndices :: Attributes -> [(Int, Attribute)]
 sortedAttributesIndices attrs = L.sortBy (\(_, Attribute name1 _) (_,Attribute name2 _) -> compare name1 name2) $ V.toList (V.indexed attrs)
 
 -- | The relation's tuple set is the body of the relation.
-newtype RelationTupleSet = RelationTupleSet { asList :: [RelationTuple] } deriving (Hashable, Show, Generic)
+newtype RelationTupleSet = RelationTupleSet { asList :: [RelationTuple] } deriving (Hashable, Show, Generic, Read)
+
+instance Read Relation where
+  readsPrec = error "relation read not supported"
 
 instance Eq RelationTupleSet where
  set1 == set2 = hset set1 == hset set2
@@ -131,7 +139,7 @@ instance Hashable RelationTuple where
       sortedTupVec = V.map (\(index, _) -> tupVec V.! index) $ V.fromList sortedAttrsIndices
   
 -- | A tuple is a set of attributes mapped to their 'Atom' values.
-data RelationTuple = RelationTuple Attributes (V.Vector Atom) deriving (Show, Generic)
+data RelationTuple = RelationTuple Attributes (V.Vector Atom) deriving (Show, Read, Generic)
 
 instance Eq RelationTuple where
   (==) tuple1@(RelationTuple attrs1 _) tuple2@(RelationTuple attrs2 _) = attributesEqual attrs1 attrs2 && atomsEqual
@@ -201,10 +209,12 @@ data RelationalExprBase a =
   --Summarize :: AtomExpr -> AttributeName -> RelationalExpr -> RelationalExpr -> RelationalExpr -- a special case of Extend
   --Evaluate relationalExpr with scoped views
   With [(WithNameExprBase a, RelationalExprBase a)] (RelationalExprBase a)
-  deriving (Show, Eq, Generic, NFData, Foldable, Functor, Traversable)
-           
+  deriving (Show, Read, Eq, Generic, NFData, Foldable, Functor, Traversable)
+
+instance Hashable RelationalExpr
+    
 data WithNameExprBase a = WithNameExpr RelVarName a
-  deriving (Show, Eq, Generic, NFData, Foldable, Functor, Traversable)
+  deriving (Show, Read, Eq, Generic, NFData, Foldable, Functor, Traversable, Hashable)
 
 type WithNameExpr = WithNameExprBase ()
 
@@ -226,15 +236,15 @@ type TypeVarName = StringType
 -- | Metadata definition for type constructors such as @data Either a b@.
 data TypeConstructorDef = ADTypeConstructorDef TypeConstructorName [TypeVarName] |
                           PrimitiveTypeConstructorDef TypeConstructorName AtomType
-                        deriving (Show, Generic, Eq, NFData)
+                        deriving (Show, Generic, Eq, NFData, Hashable, Read)
                                  
 -- | Found in data constructors and type declarations: Left (Either Int Text) | Right Int
 type TypeConstructor = TypeConstructorBase ()
 data TypeConstructorBase a = ADTypeConstructor TypeConstructorName [TypeConstructor] |
-                         PrimitiveTypeConstructor TypeConstructorName AtomType |
-                         RelationAtomTypeConstructor [AttributeExprBase a] |
-                         TypeVariable TypeVarName
-                     deriving (Show, Generic, Eq, NFData)
+                             PrimitiveTypeConstructor TypeConstructorName AtomType |
+                             RelationAtomTypeConstructor [AttributeExprBase a] |
+                             TypeVariable TypeVarName
+                           deriving (Show, Generic, Eq, NFData, Hashable, Read)
             
 type TypeConstructorMapping = [(TypeConstructorDef, DataConstructorDefs)]
 
@@ -244,13 +254,13 @@ type DataConstructorName = StringType
 type AtomTypeName = StringType
 
 -- | Used to define a data constructor in a type constructor context such as @Left a | Right b@
-data DataConstructorDef = DataConstructorDef DataConstructorName [DataConstructorDefArg] deriving (Eq, Show, Generic, NFData)
+data DataConstructorDef = DataConstructorDef DataConstructorName [DataConstructorDefArg] deriving (Eq, Show, Generic, NFData, Hashable, Read)
 
 type DataConstructorDefs = [DataConstructorDef]
 
 data DataConstructorDefArg = DataConstructorDefTypeConstructorArg TypeConstructor | 
                              DataConstructorDefTypeVarNameArg TypeVarName
-                           deriving (Show, Generic, Eq, NFData)
+                           deriving (Show, Generic, Eq, NFData, Hashable, Read)
                                     
 type InclusionDependencies = M.Map IncDepName InclusionDependency
 type RelationVariables = M.Map RelVarName GraphRefRelationalExpr
@@ -295,7 +305,7 @@ data DatabaseContext = DatabaseContext {
 type IncDepName = StringType             
 
 -- | Inclusion dependencies represent every possible database constraint. Constraints enforce specific, arbitrarily-complex rules to which the database context's relation variables must adhere unconditionally.
-data InclusionDependency = InclusionDependency RelationalExpr RelationalExpr deriving (Show, Eq, Generic, NFData)
+data InclusionDependency = InclusionDependency RelationalExpr RelationalExpr deriving (Show, Eq, Generic, NFData, Hashable, Read)
 
 type AttributeNameAtomExprMap = M.Map AttributeName AtomExpr
 
@@ -303,6 +313,8 @@ type AttributeNameAtomExprMap = M.Map AttributeName AtomExpr
 type DatabaseContextExprName = StringType
 
 type DatabaseContextExpr = DatabaseContextExprBase ()
+
+instance Hashable DatabaseContextExpr 
 
 type GraphRefDatabaseContextExpr = DatabaseContextExprBase GraphRefTransactionMarker
 
@@ -333,7 +345,10 @@ data DatabaseContextExprBase a =
   ExecuteDatabaseContextFunction DatabaseContextFunctionName [AtomExprBase a] |
   
   MultipleExpr [DatabaseContextExprBase a]
-  deriving (Show, Eq, Generic)
+  deriving (Show, Read, Eq, Generic, NFData)
+
+instance Hashable (M.Map AttributeName AtomExpr) where
+  hashWithSalt salt m = salt `hashWithSalt` M.toList m
 
 type ObjModuleName = StringType
 type ObjFunctionName = StringType
@@ -353,6 +368,8 @@ type DatabaseContextIOExpr = DatabaseContextIOExprBase ()
 
 type RestrictionPredicateExpr = RestrictionPredicateExprBase ()
 
+instance Hashable RestrictionPredicateExpr
+
 type GraphRefRestrictionPredicateExpr = RestrictionPredicateExprBase GraphRefTransactionMarker
 
 -- | Restriction predicates are boolean algebra components which, when composed, indicate whether or not a tuple should be retained during a restriction (filtering) operation.
@@ -364,7 +381,7 @@ data RestrictionPredicateExprBase a =
   RelationalExprPredicate (RelationalExprBase a) | --type must be same as true and false relations (no attributes)
   AtomExprPredicate (AtomExprBase a) | --atom must evaluate to boolean
   AttributeEqualityPredicate AttributeName (AtomExprBase a) -- relationalexpr must result in relation with single tuple
-  deriving (Show, Eq, Generic, NFData, Foldable, Functor, Traversable)
+  deriving (Show, Read, Eq, Generic, NFData, Foldable, Functor, Traversable)
 
 -- child + parent links
 -- | A transaction graph's head name references the leaves of the transaction graph and can be used during session creation to indicate at which point in the graph commits should persist.
@@ -425,6 +442,8 @@ instance Ord Transaction where
 
 type AtomExpr = AtomExprBase ()
 
+instance Hashable AtomExpr
+
 type GraphRefAtomExpr = AtomExprBase GraphRefTransactionMarker
 
 -- | An atom expression represents an action to take when extending a relation or when statically defining a relation or a new tuple.
@@ -433,13 +452,16 @@ data AtomExprBase a = AttributeAtomExpr AttributeName |
                       FunctionAtomExpr AtomFunctionName [AtomExprBase a] a |
                       RelationAtomExpr (RelationalExprBase a) |
                       ConstructedAtomExpr DataConstructorName [AtomExprBase a] a
-                    deriving (Eq,Show,Generic, NFData, Foldable, Functor, Traversable)
+                    deriving (Eq, Show, Read, Generic, NFData, Foldable, Functor, Traversable)
                        
 -- | Used in tuple creation when creating a relation.
 data ExtendTupleExprBase a = AttributeExtendTupleExpr AttributeName (AtomExprBase a)
-                     deriving (Show, Eq, Generic, NFData, Foldable, Functor, Traversable)
+                     deriving (Show, Read, Eq, Generic, NFData, Foldable, Functor, Traversable)
                               
 type ExtendTupleExpr = ExtendTupleExprBase ()
+
+instance Hashable ExtendTupleExpr
+  
 type GraphRefExtendTupleExpr = ExtendTupleExprBase GraphRefTransactionMarker
 
 --enumerates the list of functions available to be run as part of tuple expressions           
@@ -487,7 +509,12 @@ data AttributeNamesBase a = AttributeNames (S.Set AttributeName) |
                             UnionAttributeNames (AttributeNamesBase a) (AttributeNamesBase a) |
                             IntersectAttributeNames (AttributeNamesBase a) (AttributeNamesBase a) |
                             RelationalExprAttributeNames (RelationalExprBase a) -- use attribute names from the relational expression's type
-                      deriving (Eq, Show, Generic, NFData, Foldable, Functor, Traversable)
+                      deriving (Eq, Show, Read, Generic, NFData, Foldable, Functor, Traversable)
+
+instance Hashable AttributeNames
+
+instance Hashable (S.Set AttributeName) where
+  hashWithSalt salt s = salt `hashWithSalt` S.toList s
                                
 type AttributeNames = AttributeNamesBase ()
 
@@ -505,18 +532,22 @@ type GraphRefAttributeExpr = AttributeExprBase GraphRefTransactionMarker
 -- | Create attributes dynamically.
 data AttributeExprBase a = AttributeAndTypeNameExpr AttributeName TypeConstructor a |
                            NakedAttributeExpr Attribute
-                         deriving (Eq, Show, Generic, NFData, Foldable, Functor, Traversable)
+                         deriving (Eq, Show, Read, Generic, NFData, Foldable, Functor, Traversable, Hashable)
                               
 -- | Dynamically create a tuple from attribute names and 'AtomExpr's.
 newtype TupleExprBase a = TupleExpr (M.Map AttributeName (AtomExprBase a))
-                 deriving (Eq, Show, Generic, NFData, Foldable, Functor, Traversable)
-                          
+                 deriving (Eq, Show, Read, Generic, NFData, Foldable, Functor, Traversable)
+
+instance Hashable TupleExpr
+
 type TupleExpr = TupleExprBase ()
 
 type GraphRefTupleExpr = TupleExprBase GraphRefTransactionMarker
 
 data TupleExprsBase a = TupleExprs a [TupleExprBase a]
-  deriving (Eq, Show, Generic, NFData, Foldable, Functor, Traversable)
+  deriving (Eq, Show, Read, Generic, NFData, Foldable, Functor, Traversable)
+
+instance Hashable TupleExprs
 
 type GraphRefTupleExprs = TupleExprsBase GraphRefTransactionMarker
 
@@ -566,6 +597,7 @@ attrTypeVars (Attribute _ aType) = case aType of
   DateTimeAtomType -> S.empty
   ByteStringAtomType -> S.empty
   BoolAtomType -> S.empty
+  RelationalExprAtomType -> S.empty
   (RelationAtomType attrs) -> S.unions (map attrTypeVars (V.toList attrs))
   (ConstructedAtomType _ tvMap) -> M.keysSet tvMap
   (TypeVariableType nam) -> S.singleton nam
@@ -589,6 +621,7 @@ atomTypeVars DayAtomType = S.empty
 atomTypeVars DateTimeAtomType = S.empty
 atomTypeVars ByteStringAtomType = S.empty
 atomTypeVars BoolAtomType = S.empty
+atomTypeVars RelationalExprAtomType = S.empty
 atomTypeVars (RelationAtomType attrs) = S.unions (map attrTypeVars (V.toList attrs))
 atomTypeVars (ConstructedAtomType _ tvMap) = M.keysSet tvMap
 atomTypeVars (TypeVariableType nam) = S.singleton nam